small fixes and merge
