+++ client/components/FriendsScreen.tsx
@@ -1,26 +1,17 @@
 import React, { useState } from 'react';
-import {
-    FlatList,
-    Image,
-    Modal,
-    Pressable,
-    StyleSheet,
-    View,
-} from 'react-native';
+import { FlatList, Image, Pressable, StyleSheet, View } from 'react-native';
 import Button from '@/components/common/Button';
 import { Text } from '@/components/common/Text';
 import { theme } from '@/theme/theme';
 import { logoutUser } from '@/services/auth';
 import { useMutation } from '@tanstack/react-query';
 import { router } from 'expo-router';
-import MoreIcon from './forms/MoreIcon';
-import FriendModal from './forms/FriendModal';
 import { mockOthers } from '../misc/mockOthers';
 
 export default function FriendsScreen() {
     const [hovered, setHovered] = useState<string | null>(null);
     const [pressed, setPressed] = useState<string | null>(null);
-    const [showModal, setShowModal] = useState(false);
+
     const { mutate: logout } = useMutation({
         mutationFn: logoutUser,
         onSuccess: () => {
@@ -53,25 +44,6 @@ export default function FriendsScreen() {
                 <Image style={styles.avatar} source={{ uri: item.avatar }} />
                 <Text>{item.name}</Text>
             </View>
-            {(hovered === item.id || pressed === item.id) && (
-                <MoreIcon setShowModal={setShowModal} />
-            )}
-            <Pressable onPress={() => setShowModal(false)}>
-                <Modal
-                    animationType="fade"
-                    visible={showModal}
-                    transparent
-                    onRequestClose={() => setShowModal(false)}
-                >
-                    <View style={styles.modalContainer}>
-                        <Pressable style={styles.border}>
-                            <View style={styles.modalBox}>
-                                <FriendModal />
-                            </View>
-                        </Pressable>
-                    </View>
-                </Modal>
-            </Pressable>
         </Pressable>
     );
 
@@ -136,20 +108,5 @@ const styles = StyleSheet.create({
     },
     logoutButton: {
         margin: theme.spacing.sm,
-    },
-    modalContainer: {
-        flex: 1,
-        justifyContent: 'center',
-        alignItems: 'center',
-    },
-    modalBox: {
-        backgroundColor: theme.colors.background,
-    },
-    border: {
-        borderWidth: 1,
-        borderColor: theme.colors.secondary,
-        borderRadius: 25,
-        padding: theme.spacing.md,
-        cursor: 'auto',
     },
 });
